- fixed a bug in the Wind class - added .close() to realm instances that run on threads that aren't the main one - the main thread's realm instance will only be closed when the app is closed - removing data is not done in a stream, but rather in the main thread - made BaseRepository class abstract

+++ app/src/main/java/com/barbasdev/realmexample/datamodel/Wind.java
@@ -10,7 +10,7 @@ public class Wind extends RealmObject {
     private Float speed;
 
     @SerializedName("deg")
-    private Long deg;
+    private Float deg;
 
     public Float getSpeed() {
         return speed;
@@ -20,11 +20,11 @@ public class Wind extends RealmObject {
         this.speed = speed;
     }
 
-    public Long getDeg() {
+    public Float getDeg() {
         return deg;
     }
 
-    public void setDeg(Long deg) {
+    public void setDeg(Float deg) {
         this.deg = deg;
     }
